# ui/client/shared/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateBeatRequest": {
        "description": "Request body for creating a beat.\n\nNo note/tags: the app takes no free-text input. Tags are auto-derived\nserver-side from the daemon's flow signals (see BeatService).",
        "properties": {
          "end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "project_id"
        ],
        "title": "CreateBeatRequest",
        "type": "object"
      },
...
      "DeviceHeartbeatRequest": {
        "properties": {
          "battery_voltage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Voltage"
          },
          "uptime_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uptime Seconds"
          },
          "wifi_rssi": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wifi Rssi"
          }
        },
        "title": "DeviceHeartbeatRequest",
        "type": "object"
      },
      "DeviceHeartbeatResponse": {
        "properties": {
          "battery_voltage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Voltage"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "uptime_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uptime Seconds"
          },
          "wifi_rssi": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wifi Rssi"
          }
        },
        "title": "DeviceHeartbeatResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeatmapDayResponse": {
        "description": "Response schema for a single day in the contribution heatmap.",
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "project_count": {
            "title": "Project Count",
            "type": "integer"
          },
          "session_count": {
            "title": "Session Count",
            "type": "integer"
          },
          "total_minutes": {
            "title": "Total Minutes",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "total_minutes",
          "session_count",
          "project_count"
        ],
        "title": "HeatmapDayResponse",
        "type": "object"
      },
...
      "PairExchangeRequest": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          }
        },
        "required": [
          "code"
        ],
        "title": "PairExchangeRequest",
        "type": "object"
      },
      "PairExchangeResponse": {
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "device_token": {
            "title": "Device Token",
            "type": "string"
          }
        },
        "required": [
          "device_token",
          "device_id"
        ],
        "title": "PairExchangeResponse",
        "type": "object"
      },
...
      "UpdateBeatRequest": {
        "description": "Request body for updating a beat. Tags are auto-derived, not sent.",
        "properties": {
          "end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "start"
        ],
        "title": "UpdateBeatRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A time tracking application",
    "title": "Beats API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analytics/heatmap": {
      "get": {
        "description": "Get daily activity heatmap for a given year.",
        "operationId": "get_heatmap_api_analytics_heatmap_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "title": "Year",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
            "in": "query",
            "name": "tz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
              "title": "Tz"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HeatmapDayResponse"
                  },
                  "title": "Response Get Heatmap Api Analytics Heatmap Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Heatmap",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/beats/": {
      "get": {
        "description": "List beats with optional filters.",
        "operationId": "list_beats_api_beats__get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "date_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Beats",
        "tags": [
          "Beats"
        ]
      },
      "post": {
        "description": "Create a new beat (time log entry).",
        "operationId": "create_beat_api_beats__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Beat",
        "tags": [
          "Beats"
        ]
      },
      "put": {
        "description": "Update an existing beat.",
        "operationId": "update_beat_api_beats__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Beat",
        "tags": [
          "Beats"
        ]
      }
    },
...
    "/api/device/heartbeat": {
      "get": {
        "description": "Get the last heartbeat from the wall clock device.",
        "operationId": "get_heartbeat_api_device_heartbeat_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DeviceHeartbeatResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Heartbeat Api Device Heartbeat Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Heartbeat",
        "tags": [
          "device"
        ]
      },
      "post": {
        "description": "Receive a heartbeat from the wall clock device.",
        "operationId": "post_heartbeat_api_device_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceHeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceHeartbeatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Heartbeat",
        "tags": [
          "device"
        ]
      }
    },
...
    "/api/device/pair/exchange": {
      "post": {
        "description": "Exchange a pairing code for a long-lived device token (public endpoint).\n\nRate-limited because the endpoint is unauthenticated and the pairing\ncode is only ~30 bits of entropy (6 base32 chars). Without a limit,\na peer on the same NAT could grind through the keyspace.",
        "operationId": "exchange_pair_code_api_device_pair_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairExchangeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange Pair Code",
        "tags": [
          "device"
        ]
      }
    },
...
    "/api/device/status": {
      "get": {
        "description": "Get timer state optimized for ESP32 wall clock firmware.",
        "operationId": "get_device_status_api_device_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "theme",
            "required": false,
            "schema": {
              "default": "ember",
              "title": "Theme",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Device Status",
        "tags": [
          "device"
        ]
      }
    },
...
    "/api/projects/": {
      "get": {
        "description": "List projects, optionally augmented with aggregation fields.\n\nBackward compatible: without `include`, the response carries the same\nfields as ProjectResponse plus null aggregation slots \u2014 older clients\nthat ignore unknown null fields keep working.\n\nP3.0 of the project-management revamp. First cut loops per-project\nserver-side; if profiling shows it's needed, a follow-up (P3.0b)\nmoves to a Mongo $facet pipeline for true batched aggregation.",
        "operationId": "list_projects_api_projects__get",
        "parameters": [
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          },
          {
            "description": "Comma-separated aggregations to populate per project. Supported: 'totals' (total_minutes), 'this_week' (weekly_minutes + effective_goal trio, and on a day job the contract's expected / worked / remaining / balance), 'last_tracked' (last_tracked_at). Omitted \u21d2 slim response.",
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated aggregations to populate per project. Supported: 'totals' (total_minutes), 'this_week' (weekly_minutes + effective_goal trio, and on a day job the contract's expected / worked / remaining / balance), 'last_tracked' (last_tracked_at). Omitted \u21d2 slim response.",
              "title": "Include"
            }
          },
          {
            "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
            "in": "query",
            "name": "tz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
              "title": "Tz"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectsListItemResponse"
                  },
                  "title": "Response List Projects Api Projects  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Projects",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "description": "Create a new project.",
        "operationId": "create_project_api_projects__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "Projects"
        ]
      },
      "put": {
        "description": "Update an existing project.",
        "operationId": "update_project_api_projects__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "Projects"
        ]
      }
    },
...
  }
}